+++ README.md
@@ -62,7 +62,7 @@ Also look at the [jupyter notebooks](https://github.com/unit8co/vegans/tree/mast
 
 ```python
 from vegans.GAN import VanillaGAN
-import vegans.utils.utils as utils
+import vegans.utils as utils
 import vegans.utils.loading as loading
 
 # Data preparation (Load your own data or example MNIST)
@@ -100,7 +100,7 @@ utils.plot_images(samples)
 ```python
 import torch
 import numpy as np
-import vegans.utils.utils as utils
+import vegans.utils as utils
 import vegans.utils.loading as loading
 from vegans.GAN import ConditionalVanillaGAN
