+++ tutorials/snippets/example_README.py
@@ -2,7 +2,7 @@ mode = "unsupervised"
 
 if mode == "unsupervised":
     from vegans.GAN import VanillaGAN
-    import vegans.utils.utils as utils
+    import vegans.utils as utils
     import vegans.utils.loading as loading
 
     # Data preparation (Load your own data or example MNIST)
@@ -37,7 +37,7 @@ if mode == "unsupervised":
 elif mode == "supervised":
     import torch
     import numpy as np
-    import vegans.utils.utils as utils
+    import vegans.utils as utils
     import vegans.utils.loading as loading
     from vegans.GAN import ConditionalVanillaGAN
