+++ tutorials/snippets/example_input_formats.py
@@ -1,5 +1,5 @@
 import numpy as np
-import vegans.utils.utils as utils
+import vegans.utils as utils
 import vegans.utils.loading as loading
 
 from torch import nn
