+++ tutorials/snippets/example_mnist_rotation.py
@@ -1,6 +1,6 @@
 import numpy as np
 import matplotlib.pyplot as plt
-import vegans.utils.utils as utils
+import vegans.utils as utils
 import vegans.utils.loading as loading
 
 from vegans.GAN import (
