+++ vegans/models/unconditional/WassersteinGAN.py
@@ -27,7 +27,7 @@ import torch
 import numpy as np
 
 from vegans.models.unconditional.AbstractGAN1v1 import AbstractGAN1v1
-from vegans.utils.utils import WassersteinLoss
+from vegans.utils import WassersteinLoss
 
 
 class WassersteinGAN(AbstractGAN1v1):
