+++ vegans/utils/__init__.py
@@ -1,0 +1,3 @@
+from vegans.utils.losses import *
+from vegans.utils.plotting import *
+from vegans.utils.processing import *
