+++ vegans/utils/loading/architectures/example.py
@@ -1,7 +1,7 @@
 import numpy as np
 import torch.nn as nn
 
-from vegans.utils.utils import get_input_dim
+from vegans.utils import get_input_dim
 from vegans.utils.layers import LayerReshape
 
 class MyGenerator(nn.Module):
