+++ vegans/utils/plot2DModel.py
@@ -4,7 +4,7 @@ import torch
 import numpy as np
 import pandas as pd
 import matplotlib as mpl
-import vegans.utils.utils as utils
+import vegans.utils as utils
 import matplotlib.pyplot as plt
